Deprecate top level classes
